+++ src/Markdown.ColorCode/Internal/CodeExtractor.cs
@@ -1,0 +1,35 @@
+﻿namespace Markdown.ColorCode.Internal;
+
+/// <inheritdoc cref="ICodeExtractor"/>
+internal sealed class CodeExtractor : ICodeExtractor
+{
+    /// <inheritdoc />
+    public string ExtractCode(LeafBlock leafBlock)
+    {
+        var code = new StringBuilder();
+        var lines = leafBlock.Lines.Lines ?? Array.Empty<StringLine>();
+        var totalLines = lines.Length;
+
+        for (var index = 0; index < totalLines; index++)
+        {
+            var line = lines[index];
+            var slice = line.Slice;
+
+            if (slice.Text == null)
+            {
+                continue;
+            }
+
+            var lineText = slice.Text.Substring(slice.Start, slice.Length);
+
+            if (index > 0)
+            {
+                code.AppendLine();
+            }
+
+            code.Append(lineText);
+        }
+
+        return code.ToString();
+    }
+}