+++ tests/Markdown.ColorCode.UnitTests/AssemblyInfo.cs
@@ -1,0 +1,13 @@
+﻿global using ColorCode;
+global using ColorCode.Styling;
+global using FluentAssertions;
+global using HtmlAgilityPack;
+global using Markdig;
+global using Markdig.Helpers;
+global using Markdig.Parsers;
+global using Markdig.Renderers;
+global using Markdig.Renderers.Html;
+global using Markdig.Syntax;
+global using Markdown.ColorCode.Internal;
+global using NSubstitute;
+global using NUnit.Framework;